Back ported [PD-985] fix from my pd-961 branch
